incorporates new Fuel UX Handlebars templates and utilizes them for Scheduler example
